requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
passlib[bcrypt]  
pydantic[email]
python-jose[cryptography]  
jwt
jinja2
authlib
python-dotenv
aiohttp
pydantic[extra]
requests
starlette
feedparser
transformers
python-multipart
beautifulsoup4
